add schemas index
